python/requirements.txt
# --------- base-tool --------- #
sphinx
sphinx-rtd-theme
pytest>=7.4.3
pytest-cov>=4.1.0
pyrootutils>=1.0.4          # standardizing the project root setup
pre-commit>=3.5.0
twine>=4.0.1
wheel>=0.41.3
coverage>=7.3.2
furo
myst-parser
pytest-mock
pendulum
# --------- date --------- #
pendulum>=3.0.0  # pendulum>=3.0.0 对时间 pendulum.parse('2023-01-01') 处理不一样,3.0.0更灵活
openpyxl
attrs

# --------- path --------- #
pathlib


# --------- web --------- #
streamlit


# --------- cli --------- #


# --------- log --------- #
loguru
prettytable

# --------- ai-tools --------- #
langchain-openai
langchain-community
langchain_cohere
langchain_postgres
openai
duckduckgo-search
whisper
arxiv



# --------- server --------- #
fastapi
uvicorn


# --------- db --------- #
psycopg[binary]
langchain-chroma
chromadb
# faiss-cpu


# --------- deploy --------- #
# ansible

# --------- data --------- #
pydantic
# numpy==1.26.4
numpy
pandas
# polars
pypdf
pyarrow
unstructured
python-docx
python-magic
python-pptx
docx2txt
markdown

# --------- graph database --------- #
neo4j
# neomodel[extras]


# --------- yaml --------- #
pyyaml

# --------- agent --------- #
crewai[tools]
# pyautogen
dspy-ai
mem0ai


# --------- distributed --------- #
dora-rs

# --------- stock --------- #
yfinance

# --------- embedding --------- #
sentence_transformers
transformers





